pyproject.toml
[project]
name = "beatpulse"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "requests>=2.32.4",
    "requests-oauthlib>=2.0.0",
]
